Add files via upload
A Markdown previewer is a tool or application that allows users to input Markdown syntax, which is a lightweight markup language, and see the formatted output in real-time. Markdown is commonly used for writing content for the web, such as blog posts, documentation, and readme files, as it provides a simple and intuitive way to format text without needing to write HTML directly.